Dev user can seed database
the XML we're parsing is a bit gnarly. It has a few variations on addresses and
grants and org names/ein, so after accounting for as many of those variations as
I could find, I think this script is robust enough for me to move on.

Also found the `Array |= []` syntax helpful for making sure to only add a record
to a collection once.